make deltasec a float type

+++ plugins/parking/src/parking.h
@@ -65,7 +65,7 @@ class XletParking : public XLet
     private:
         QVBoxLayout * m_parkinglayout;
         int m_timerid;  //!< id of the timer
-        int m_deltasec;  //!< timer period
+        double m_deltasec;  //!< timer period
         // QList<PeerChannel *> m_mychannels;  //!< "my channels" list for transfer menu
         QHash<QString, ParkingWidget *> m_parkinglots; //!< Parkinglots widgets
 };
